+++ .gitignore
@@ -1,11 +1,12 @@
 /node_modules/
-/dist/
+/src/dist/
 *.log
 play.js
-/public/*
-!/public/.keep
-!/public/index.html
-!/public/index-dev.html
-!/public/index-idle.html
+/src/public/*
+!/src/public/.keep
+!/src/public/index.html
+!/src/public/index-dev.html
+!/src/public/index-idle.html
+!/src/public/images
 t1.js
 dist-pkg
