+++ package.json
@@ -2,21 +2,21 @@
   "name": "record-desktop",
   "version": "0.5.5",
   "description": "Record gifs and take screenshots on linux, built with electron.",
-  "main": "dist/main.js",
+  "main": "src/dist/main.js",
   "scripts": {
-    "start": "NODE_ENV=development electron ./dist/main.js",
+    "start": "NODE_ENV=development electron ./src/dist/main.js",
     "test": "ava 'src/**/*.spec.js'",
     "test:watch": "nodemon --exec ava",
     "prebuild": "rm -rf dist && mkdir dist",
     "build": "npm-run-all --parallel build:electron build:frontend",
     "build:watch": "npm-run-all --parallel build:electron:watch build:frontend:watch",
-    "build:electron": "babel -d dist/ src/",
-    "build:electron:watch": "babel --watch -d dist/ src/",
+    "build:electron": "babel -d src/dist src/main",
+    "build:electron:watch": "babel --watch -d src/dist src/main",
     "build:frontend": "NODE_ENV=production webpack",
     "build:frontend:watch": "webpack-dev-server",
     "precommit": "npm run lint",
     "electron": "NODE_ENV=development electron ./dist/main.js",
-    "lint": "eslint src/* frontend/**/*.js",
+    "lint": "eslint 'src/main/**/*.js' 'src/frontend/**/*.js'",
     "build:electron-pkg": "npm run build && rm -rf dist-pkg && electron-packager . record-desktop --platform linux --arch x64 --out dist-pkg/",
     "build:electron-deb": "npm run build:electron-pkg && electron-installer-debian --src dist-pkg/record-desktop-linux-x64/ --dest dist-pkg/installers/ --arch amd64"
   },
