+++ src/main/config.js
