+++ src/main/logger.js
