+++ src/main/shortcuts.js
