+++ src/main/unix-utils/wrappers/byzanz-record.js
