+++ src/main/unix-utils/wrappers/mkfifo.js
@@ -6,7 +6,8 @@ export default function mkfifo() {
   const file = `/tmp/fifo-${Math.random()}`;
   execSync(`rm -f ${file} && mkfifo ${file}`);
 
-  const proc = spawnAsync(`sh`, [path.resolve(__dirname, '..', '..', '..', 'cat.sh'), file]);
+  // @TODO why
+  const proc = spawnAsync(`sh`, [path.resolve(__dirname, '..', '..', '..', '..', 'cat.sh'), file]);
   const onEnd = () => proc.kill();
 
   proc.stdout.pipe(process.stdout);
