+++ src/main/unix-utils/wrappers/slop.spec.js
