+++ src/main/unix-utils/wrappers/xwd.js
