+++ src/main/unix-utils/wrappers/xwininfo.js
