+++ src/public/images/icon-512x512-stroke.png
