+++ src/renderer/components/gallery-file/gallery-file.js
