+++ src/renderer/components/gallery-file/style.css
