+++ src/renderer/components/gallery/detect-viewport.js
