+++ src/renderer/components/gallery/gallery.js
