+++ src/renderer/components/settings/folder-select.js
