+++ src/renderer/components/settings/has-notifications-input.js
