+++ src/renderer/components/settings/settings.js
