+++ src/renderer/components/settings/shadow-select.js
