+++ src/renderer/components/settings/shortcut-input.js
