+++ src/renderer/components/settings/style.css
